+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Requires PHP `8.4`
 - Requires `innmind/server-control:~7.0`
+- Requires `innmind/time:~1.0`
 
 ## 5.0.0 - 2025-05-08